% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Validation d'un sudoku}
\usage{
check(A)
}
\arguments{
\item{A}{Sudoku (Matrice  9x9)}
}
\value{
TRUE si le sudoku est valide, FALSE sinon.
}
\description{
Cette fonction verifie à la fin de la partie le sudoku est valide ou non
(pas de nombre qui se repete dans une ligne, colonne ou bloc 3x3).
}
\examples{
check(sudo_ele())
}
\seealso{
Du package MALEX : is_valid
}
