% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{game}
\alias{game}
\title{A partir d'une difficulte, retourne un sudoku correspondant avec sa solution.}
\usage{
game(diff)
}
\arguments{
\item{diff, }{Une chaine de caractere appartenant a "Facile", "Moyen", "Difficile", "Cauchemar"}
}
\value{
une liste contenant : un sudoku de difficulte diff, sa solution, et le nombre de backtracking necessaire a l'algorithme pour le resoudre.
}
\description{
A partir d'une difficulte, retourne un sudoku correspondant avec sa solution.
}
