% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sudo.R
\name{get_sudo}
\alias{get_sudo}
\title{Genere un sudoku aleatoire a partir d'un sudoku elementaire.}
\usage{
get_sudo()
}
\value{
Une matrice 9x9 correspondant a un sudoku elementaire
}
\description{
Genere un sudoku aleatoire a partir d'un sudoku elementaire.
}
