% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_bloc.R
\name{in_bloc}
\alias{in_bloc}
\title{Verification par bloc}
\usage{
in_bloc(A, nrow, ncol, num)
}
\arguments{
\item{A}{Matrice 9X9 sous forme d'un sudoku}

\item{nrow}{Numero de la ligne}

\item{ncol}{Numero de la colonne}

\item{num}{Un chiffre de 1 a 9}
}
\value{
Renvoie True si le chiffre existe deja dans le bloc 3x3 et False sinon.
}
\description{
Cette fonction verifie si un chiffre existe deja dans un bloc du
sudoku.
}
\examples{
in_bloc(sudo_ele(),1,1,6)
}
