% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_col.R
\name{in_col}
\alias{in_col}
\title{Verification par colonne}
\usage{
in_col(A, ncol, num)
}
\arguments{
\item{A, }{Matrice}

\item{ncol, }{numero de la colonne}

\item{num, }{un chiffre de 1 a 9}
}
\value{
True si le chiffre existe deja dans la colonne et False sinon.
}
\description{
Cette fonction verifie si un chiffre existe deja dans une colonne
du sudoku.
}
\examples{
in_col(sudo_ele(),1,8)
}
