% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_line.R
\name{in_line}
\alias{in_line}
\title{Verification par ligne}
\usage{
in_line(A, nrow, num)
}
\arguments{
\item{A}{Matrice  9x9}

\item{nrow}{numero de la ligne}

\item{num}{un chiffre de 1 a 9}
}
\value{
True si le chiffre existe deja dans la ligne et False sinon.
}
\description{
Cette fonction verifie si un chiffre existe deja dans une ligne
du sudoku.
}
\examples{
in_line(sudo_ele(),1,6)
}
