% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_possible.R
\name{is_possible}
\alias{is_possible}
\title{Verification globale}
\usage{
is_possible(A, nrow, ncol, num)
}
\arguments{
\item{A}{Matrice  9x9}

\item{nrow}{numero de la ligne}

\item{ncol}{numero de la colonne}

\item{num}{un chiffre de 1 a 9}
}
\value{
TRUE si l'emplacement est possible et FALSE sinon
}
\description{
Cette fonction verifie si l'emplacement du chiffre dans un
sudoku est possible. Cette dernière utilise les differentes fonctions de
verification du package.
}
\examples{
is_possible(play_sudo(sudo_ele(),50),1,1,4)
}
\seealso{
Du package MALEX : in_line, in_col, in_bloc.
}
