% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid}
\alias{is_valid}
\title{Cette fonction elle verifie si un sudoku est valide ou pas
(pas de nombre qui se repete dans une ligne, colonne ou bloc 3x3)}
\usage{
is_valid(A, nrow, ncol, num)
}
\arguments{
\item{A}{Sudoku (Matrice  9x9)}

\item{nrow}{Numero de la ligne}

\item{ncol}{Numero de la colonne}

\item{num}{Un chiffre de 1 a 9}
}
\value{
TRUE si le sudoku est valide, FALSE sinon.
}
\description{
Cette fonction elle verifie si un sudoku est valide ou pas
(pas de nombre qui se repete dans une ligne, colonne ou bloc 3x3)
}
