% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permu_aleat.R
\name{permu_aleat}
\alias{permu_aleat}
\title{Permute aléatoirement une matrice 9x9 par bloc (3x9 et 9x3)}
\usage{
permu_aleat(A)
}
\arguments{
\item{Matrice}{9x9}
}
\value{
Une matrice 9x9 permuter
}
\description{
Permute aléatoirement une matrice 9x9 par bloc (3x9 et 9x3)
}
