% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sudoku.R
\name{plot_sudoku}
\alias{plot_sudoku}
\title{Affichage d'un sudoku}
\usage{
plot_sudoku(A)
}
\arguments{
\item{A}{Matrice 9x9}
}
\value{
Affichage graphique du sudoku.
}
\description{
Cette fonction permet d'afficher un sudoku sous sa forme
standard.
}
