% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_case.R
\name{solve_case}
\alias{solve_case}
\title{Cette fonction résolue une case d'un sudoku}
\usage{
solve_case(A, nrow, ncol)
}
\arguments{
\item{A}{Matrice  9x9}

\item{nrow}{numéro de la ligne}

\item{ncol}{numéro de la colonne}
}
\value{
True si la case est résolue False sinon.
}
\description{
Cette fonction résolue une case d'un sudoku
}
