% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_sudo.R
\name{solve_sudo}
\alias{solve_sudo}
\title{Resolue le sudoku par la methode du backtracking}
\usage{
solve_sudo(sudoku)
}
\arguments{
\item{sudoku}{Matrice  9x9 (avec des valeurs manquantes NA et non pas des zeros)}
}
\value{
Liste qui contient le sudoku resolu et le nombre de retour que l'algorithme a fait pour le resoudre
}
\description{
Resolue le sudoku par la methode du backtracking
}
