% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudo_ele.R
\name{sudo_ele}
\alias{sudo_ele}
\title{Produit un sudoku élémentaire de manière aléatoire}
\usage{
sudo_ele()
}
\value{
Une matrice 9x9 correspondant à un sudoku élémentaire
}
\description{
Produit un sudoku élémentaire de manière aléatoire
}
