% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfo.R
\name{transfo}
\alias{transfo}
\title{Transformation de matrice}
\usage{
transfo(A)
}
\arguments{
\item{A}{Matrice  9x9}
}
\value{
Matrice 9x9 avec des zeros au lieu des NA.
}
\description{
Cette fonction transforme une matrice 9x9 avec des NA en matrice
 9x9 avec des zeros a la place.
}
\examples{
A <- play_sudo(sudo_ele(), 45)
print(transfo(A))
}
