% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{game}
\alias{game}
\title{A partir d'une difficulte, retourne un sudoku correspondant avec sa solution.}
\usage{
game(diff)
}
\arguments{
\item{diff, }{Une chaine de caractere appartenant a c("Facile", "Moyen", "Difficile", "Cauchemar")}
}
\value{
une liste contenant : un sudoku de difficulté diff, sa solution, et le nombre de backtracking nécessaire à l'algorithme pour le résoudre.
}
\description{
A partir d'une difficulte, retourne un sudoku correspondant avec sa solution.
}
