% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possibility.R
\name{get_possibility}
\alias{get_possibility}
\title{Calcule le nombre de possibilité pour chaque case du sudoku (matrice 9x9)}
\usage{
get_possibility(A)
}
\arguments{
\item{A}{Matrice  9x9}
}
\value{
Matrice 9x9 qui contient dans chaque case le nombre de chiffre possible à mettre dans cette case.
}
\description{
Calcule le nombre de possibilité pour chaque case du sudoku (matrice 9x9)
}
