% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possibility.R
\name{get_possibility}
\alias{get_possibility}
\title{Calcule le nombre de possibilité pour chaque case du sudoku (matrice 9x9)}
\usage{
get_possibility(A)
}
\arguments{
\item{A}{Matrice  9x9}
}
\value{
Liste qui contient une matrice 9x9 (avec les cases remplis par le nombre de possiblités) et une Dataframe
}
\description{
Calcule le nombre de possibilité pour chaque case du sudoku (matrice 9x9)
}
