% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sudo.R
\name{get_sudo}
\alias{get_sudo}
\title{Génère un sudoku aléatoire à partir d'un sudoku élémentaire.}
\usage{
get_sudo()
}
\value{
Une matrice 9x9 correspondant à un sudoku élémentaire
}
\description{
Génère un sudoku aléatoire à partir d'un sudoku élémentaire.
}
