% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_bloc.R
\name{in_bloc}
\alias{in_bloc}
\title{Vérifie si un chiffre existe déjà dans un bloc 3x3 donnée}
\usage{
in_bloc(A, nrow, ncol, num)
}
\arguments{
\item{A}{Matrice}

\item{nrow}{Numéro de la ligne}

\item{ncol}{Numéro de la colonne}

\item{num}{Un chiffre de 1 à 9}
}
\value{
True si le chiffre existe déjà dans le bloc 3x3 et False sinon.
}
\description{
Vérifie si un chiffre existe déjà dans un bloc 3x3 donnée
}
