% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_possible.R
\name{is_possible}
\alias{is_possible}
\title{Verifie si l'emplacement du chiffre dans un sudoku est possible}
\usage{
is_possible(A, nrow, ncol, num)
}
\arguments{
\item{A}{Matrice  9x9}

\item{nrow}{numero de la ligne}

\item{ncol}{numero de la colonne}

\item{num}{un chiffre de 1 ? 9}
}
\value{
TRUE si l'emplacement est possible et FALSE sinon
}
\description{
Verifie si l'emplacement du chiffre dans un sudoku est possible
}
