% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid}
\alias{is_valid}
\title{Cette fonction elle vérifie si un sudoku est valide ou pas
(pas de nombre qui se répéte dans une méme ligne, colonne ou bloc 3x3)}
\usage{
is_valid(A, nrow, ncol, num)
}
\arguments{
\item{A}{Sudoku (Matrice  9x9)}

\item{nrow}{Numéro de la ligne}

\item{ncol}{Numéro de la colonne}

\item{num}{Un chiffre de 1 à 9}
}
\value{
TRUE si le sudoku est valide, FALSE sinon.
}
\description{
Cette fonction elle vérifie si un sudoku est valide ou pas
(pas de nombre qui se répéte dans une méme ligne, colonne ou bloc 3x3)
}
