% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_sudo.R
\name{play_sudo}
\alias{play_sudo}
\title{Algorithme qui masque certaines valeurs du sudoku pour commencer a jouer.}
\usage{
play_sudo(A, N)
}
\arguments{
\item{A, }{Matrice 9x9}

\item{N, }{Nombre de valeurs manquantes dans le sudoku}
}
\value{
Une matrice 9x9 de sudoku prete a jouer.
}
\description{
Algorithme qui masque certaines valeurs du sudoku pour commencer a jouer.
}
