% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_sudo.R
\name{play_sudo}
\alias{play_sudo}
\title{Algorithme qui masque certaines valeurs du sudoku pour commencer à jouer.}
\usage{
play_sudo(A)
}
\arguments{
\item{M, }{Matrice 9x9}
}
\value{
Une matrice 9x9 de sudoku prête à jouer.
}
\description{
Algorithme qui masque certaines valeurs du sudoku pour commencer à jouer.
}
