% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_sudo.R
\name{solve_sudo}
\alias{solve_sudo}
\title{Résolue le sudoku par la méthode du backtracking}
\usage{
solve_sudo(sudoku)
}
\arguments{
\item{sudoku}{Matrice  9x9 (avec des valeurs manquantes NA et non pas des zéros)}
}
\value{
Liste qui contient le sudoku résolu et le nombre de retour que l'algorithme a fait pour le résoudre
}
\description{
Résolue le sudoku par la méthode du backtracking
}
