% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfo.R
\name{transfo}
\alias{transfo}
\title{Transforme une matrice 9x9 avec des NA en matrice 9x9 avec des zeros a la place.}
\usage{
transfo(A)
}
\arguments{
\item{A}{Matrice  9x9}
}
\value{
Matrice 9x9 avec des zeros au lieu des NA
}
\description{
Transforme une matrice 9x9 avec des NA en matrice 9x9 avec des zeros a la place.
}
